+++ cookbooks/eyecandy/recipes/default.rb
@@ -39,11 +39,13 @@ if selected
   node.default[:apps] = { :eyecandy => eyecandy }
 
   include_recipe "eyecandy::greeter-background" if selected.include?("greeter-background")
-  include_recipe "eyecandy::faenza-icons" if selected.include?("faenza-icons")
-  include_recipe "eyecandy::faience-icons" if selected.include?("faience-icons")
+  include_recipe "eyecandy::faba-icons" if selected.include?("faba-icons")
   include_recipe "eyecandy::moka-icons" if selected.include?("moka-icons")
+  include_recipe "eyecandy::moka-theme" if selected.include?("moka-theme")
   include_recipe "eyecandy::nitrux-icons" if selected.include?("nitrux-icons")
-  include_recipe "eyecandy::numix-icons" if selected.include?("numix-icons")
+  include_recipe "eyecandy::numix-circle-icons" if selected.include?("numix-circle-icons")
+  include_recipe "eyecandy::numix-shine-icons" if selected.include?("numix-shine-icons")
+  include_recipe "eyecandy::orchis-theme" if selected.include?("orchis-theme")
   include_recipe "eyecandy::unsettings" if selected.include?("unsettings")
 end